Move the simpleEncryption test into the sdk test suite
