+++ libs/trustchain/src/__tests__/unit/sdk.test.ts
@@ -54,6 +54,20 @@ describe("Trustchain SDK", () => {
     Object.values(deviceCallbacks).forEach(mock => mock.mockClear());
   });
 
+  it("encryptUserData + decryptUserData", async () => {
+    const sdk = new SDK({ applicationId: 16, name: "test" }, hwDeviceProviderMock);
+    const obj = new Uint8Array([1, 2, 3, 4, 5]);
+    const keypair = await crypto.randomKeypair();
+    const trustchain = {
+      rootId: "",
+      walletSyncEncryptionKey: crypto.to_hex(keypair.privateKey),
+      applicationPath: "m/0'/16'/0'",
+    };
+    const encrypted = await sdk.encryptUserData(trustchain, obj);
+    const decrypted = await sdk.decryptUserData(trustchain, encrypted);
+    expect(decrypted).toEqual(obj);
+  });
+
   it("should create Trustchain", async () => {
     // Mock trustchain states:
     const device = new SoftwareDevice(convertLiveCredentialsToKeyPair(MOCK_DATA.members.alice));
